+++ tests/server/example/admin.py
@@ -34,7 +34,7 @@ from example.models import (
     Task,
     User,
 )
-from example.views import CrispyFormView
+from example.views import CrispyFormView, ModelExtraUrlView
 from unfold.admin import ModelAdmin, StackedInline, TabularInline
 from unfold.contrib.filters.admin import (
     AllValuesCheckboxFilter,
@@ -1097,6 +1097,15 @@ class ProjectAdmin(ModelAdmin, ImportExportModelAdmin):
     import_form_class = ImportForm
     export_form_class = SelectableFieldsExportForm
     search_fields = ["name"]
+
+    def get_custom_urls(self):
+        return [
+            (
+                "extra-url",
+                "custom_url_name",
+                ModelExtraUrlView.as_view(model_admin=self),
+            ),
+        ]
 
 
 @admin.register(Task)